user attempts logs in after not having logged out
In case the client has lost their access/refresh token (e.g. localStorage was cleared),
an attempt to log in would result in a `ValidationError` because collection RefreshToken
has user id as a unique field. To overcome this issue, we simply do nothing on such error
and continue with the regular log in process flow.